Move components from src/ to src/components/, add unstyled login page skeleton.
